ENG-16304: Disable views just before snapshot processing
During an elastic join the new partition leaders are promoted one at a time.
This means there a points in time when an mp request can get to a subset of the
new partitions. The previous invocation location of setViewsEnabled could
result in a subset of the new partitions waiting in the ee at the count down
latch. Then if another MP transaction was executed those sites would not
process the fragment task and cause an MP deadlock.

Do not call setViewsEnabled until all partitions are participating in the MP
transactions and this known when the first block of snapshot data is processed.